Working - HttpService project code reourganised and logging added fro cache hit and miss. Also Cache integrated with code.
HttpService package structure reorganised to move all HttpProcessor and Cache related functionality to a new package. Also, the cache functionality has been integrated with the HttpService code.
Logging added for cache miss and hit, also lock cache hit and miss- this is also working fine.
....Decorator class has been renamed to ...Cache.

Next step is to increase the days from 1 to 3 and see if the images pop up, cache works fine to limit the requests to Underground Weather server.